Backport fixes and token fields
  - all - Includes backports of fixes for those upgrading from pre 1.x
  - db - Backport of  backup/restore commands, saves everything but sessions
  - capture - Backport of tokens for host.*/http.uri/http.useragent field,
              can't use in viewer though until Moloch 1.8

+++ capture/writer-null.c
@@ -26,7 +26,7 @@
 
 extern MolochConfig_t        config;
 
-LOCAL  uint32_t              outputFilePos = 24;
+LOCAL  uint64_t              outputFilePos = 24;
 
 /******************************************************************************/
 LOCAL uint32_t writer_null_queue_length()
